+++ sdk/utils/configuration/__init__.py
@@ -1,6 +1,6 @@
-from .base_config import BaseConfig, InvalidConfiguration
-from .emu_free import EmuFreeConfig
-from .emu_sv import EmuSVConfig
+from sdk.utils.configuration.base_config import BaseConfig, InvalidConfiguration
+from sdk.utils.configuration.emu_free import EmuFreeConfig
+from sdk.utils.configuration.emu_sv import EmuSVConfig
 
 __all__ = [
     "BaseConfig",