+++ sdk/utils/configuration/base_config.py
@@ -15,7 +15,7 @@ INVALID_KEY_ERROR_MSG = "Invalid key {} in Configuration.extra_config. Attempted
 class BaseConfig:
     """Base class for all configuration classes.
     
-    This class provides a common interface for all configuration classes.
+    This class provides a common interface for all device configuration classes.
     It also provides a mechanism to add extra configuration parameters
     that are not part of the default configuration.