+++ sdk/utils/configuration/emu_free.py
@@ -8,7 +8,7 @@ class EmuFreeConfig(BaseConfig):
     """Configuration for the EmuFree device type.
 
     Args:
-        with_noise (bool): Whether to add noise to the measurements. Defaults to False.
+        with_noise (bool): Whether to add noise to the simulation. Defaults to False.
     """
 
     with_noise: bool = False